Rework query id Utilization (#9)
Previously we had pinned the query_id as a constant value to every instantiation of DuneAnalytics. This didn't make sense and doesn't allow for variable naming conventions. Instead we use the DuneQuery class to store the query id attribute, which is then used when passing queries into DuneAPI class (formerly known as DuneAnalytics).

This will allow for more robust features in the future (such as dashboard management).